Move top contributors function in tounesol/utils
