+++ backend/tournesol/utils/top_contributors.py
